enable multiple config files
To enable multiple config files, the importing of defaults had to
change. Now, there is a separate function `read_config_file` which
accepts a path and extracts the config as a dictionary. jarjar
configuration is extracted from expected locations (working dir; ~)
and a list of configuration is constructed in order of source
priority:

1. as kwargs to __init__
2. From a user-defined file path
3. Local .jarjar file
4. Use home ~/.jarjar file