Add screen status to attachment payload
This involves saving a couple more env variables, one for the screen
name and a flag for the intention to exit after the notification.

Now jarjar exits if such a flag was written, which means we need to
read it after the notification. Because of that screen bug, the weird
if logic needed to be copied, so I abstracted it out into a function
which creates the bash ifthen based on a check of equality.